requirements.txt
streamlit
streamlit-authenticator
pandas
numpy
yfinance
pyyaml
